SDL2 -> SDL3
SDL3 is built from source by default and optionally statically linked; both of these are controlled using Cargo features in the CLI and GUI packages

AppImage builds need to statically link SDL3 when building SDL3 from source because otherwise the AppImage build process won't locate the SDL3 shared library file (possibly a way to fix this but I didn't see anything obvious)

these changes required removing two config settings:
* borderless vs. exclusive fullscreen, since changing this seems to be much more difficult in SDL3
* audio hardware/device queue size, which SDL3 no longer seems to expose directly

+++ common/jgenesis-common/src/frontend.rs
@@ -201,8 +201,8 @@ pub trait MappableInputs<Button> {
     #[allow(unused_variables)]
     fn handle_mouse_motion(
         &mut self,
-        x: i32,
-        y: i32,
+        x: f32,
+        y: f32,
         frame_size: FrameSize,
         display_area: DisplayArea,
     ) {
